refactor: add rpc base abstract
