+++ packages/connection/src/common/proxy/index.ts
@@ -1,0 +1,11 @@
+export * from './wrapper';
+export * from './json-rpc';
+
+export abstract class RPCService<T = any> {
+  rpcClient?: T[];
+  rpcRegistered?: boolean;
+  register?(): () => Promise<T>;
+  get client() {
+    return this.rpcClient ? this.rpcClient[0] : undefined;
+  }
+}
