+++ packages/connection/src/common/utils.ts
@@ -57,3 +57,34 @@ export function getCapturer() {
   }
   return;
 }
+
+export function getServiceMethods(service: any): string[] {
+  let props: any[] = [];
+
+  if (/^\s*class/.test(service.constructor.toString())) {
+    let obj = service;
+    do {
+      props = props.concat(Object.getOwnPropertyNames(obj));
+    } while ((obj = Object.getPrototypeOf(obj)));
+    props = props.sort().filter((e, i, arr) => e !== arr[i + 1] && typeof service[e] === 'function');
+  } else {
+    for (const prop in service) {
+      if (service[prop] && typeof service[prop] === 'function') {
+        props.push(prop);
+      }
+    }
+  }
+
+  return props;
+}
+
+export function getNotificationName(serviceName: string, name: string) {
+  return `on:${serviceName}:${name}`;
+}
+export function getRequestName(serviceName: string, name: string) {
+  return `${serviceName}:${name}`;
+}
+
+export function getMethodName(serviceName: string, name: string) {
+  return name.startsWith('on') ? getNotificationName(serviceName, name) : getRequestName(serviceName, name);
+}
